fix: DispatcherTimer with default Interval on Skia and WASM
